[editorial] Assert normal completion value
The Source Text Module Record's ModuleDeclarationInstantiation method
begins by recursively resolving all requested modules of all
dependencies.

The definition of HostResolveImportedModule includes the following
invariant:

> - This operation must be idempotent if it completes normally. Each
>   time it is called with a specific referencingModule, specifier pair
>   as arguments it must return the same Module Record instance.

This means that any invocation of HostResolveImportedModule that follows
the initial traversal cannot result in an abrupt completion. Because of
this, the GetExportedNames method will itself never return an abrupt
completion.

Update the specification text to document this expectation.