[editorial] Remove unnecessary early error
Among the Early Errors for ModuleBody is:

> - It is a Syntax Error if any element of the ExportedBindings of
>   ModuleItemList does not also occur in either the VarDeclaredNames of
>   ModuleItemList, or the LexicallyDeclaredNames of ModuleItemList.

This rule, coupled with strict mode restrictions on BindingIdentifiers,
ensures that a Syntax Error will be produced when the ReferencedBinding
of a "local" export (an export matching the production `export
ExportClause ;`) is an invalid IdentifierReference. A specific early
error restricting this is not necessary.

Remove the early error and abstract operation defined to implement it.