Normative: Consistently handle abrupt completion
PerformPromiseAll creates a "Promise.all Resolve Element Function" for
each element of the array passed to `Promise.all`. It invokes each of
these indirectly through the `then` method:

>     i. Let nextPromise be ? Invoke(constructor, "resolve", « nextValue »).
>     j. Let resolveElement be a new built-in function object as defined in
>        `Promise.all` Resolve Element Functions.
>     [...]
>     q. Perform ? Invoke(nextPromise, "then", « resolveElement,
>        resultCapability.[[Reject]] »).

The Promise.all Resolve Element Function includes the logic that
ultimately fulfills the promise returned by `Promise.all`:

>     10. If remainingElementsCount.[[Value]] is 0, then
>         a. Let valuesArray be CreateArrayFromList(values).
>         b. Return ? Call(promiseCapability.[[Resolve]], undefined, «
>            valuesArray »).

But due to the way this function is invoked, if the "[[Resolve]]"
capability function returns an abrupt completion, this is ignored and
the resulting promise is never settled. This is inconsistent with the
treatment of the abrupt completion in cases where when it is handled
from PerformPromiseAll itself:

>     6.d.iii.2. Perform ? Call(resultCapability.[[Resolve]], undefined,
>                « valuesArray »).

The abrupt completion from the above call is returned to `Promise.all`,
which handles it by rejecting the promise:

>     7. Let result be PerformPromiseAll(iteratorRecord, C,
>        promiseCapability).
>     8. If result is an abrupt completion, then
>        a. If iteratorRecord.[[Done]] is false, let result be
>           IteratorClose(iterator, result).
>        b. IfAbruptRejectPromise(result, promiseCapability).

Demonstration:

    var first;
    var P = function(executor) {
      if (!first) {
        return new Promise(executor);
      }
      first = false;
      return new Promise(function(_, reject) {
        executor(function() { throw 1; }, reject);
      });
    };
    P.resolve = Promise.resolve;

    first = true;
    Promise.all.call(P, []); // this promise is rejected

    first = true;
    Promise.all.call(P, [1]); // this promise is never settled

Update the definition of "Promise.all Resolve Element Functions" to
respond to abrupt completions by invoking the "[[Reject]]" capability.